cli: suggest fixes via find_optimal_configs.py on validation failure
When collect_validation_errors returns a non-empty list, validate_args now
appends a 'Suggested fixes:' block to the ValueError message listing the
proposed CLI flags. Same proposer surface that powers find_optimal_configs.py
when run directly — users see actionable suggestions inline with the
violation report instead of having to manually re-run the diagnostic.

To avoid the circular-import dance of having cli.py reach into utils/, move
the proposer surface itself into cli.py:

- _SEARCH_RANGES, _CROSS_PARAM_FIELDS constants
- _check_cross_constraints (cross-replica constraint checker)
- _solve_cross_param_constraints (bounded grid search)
- _round_to_multiple, propose_simple_fix
- New _build_suggestion_block helper that formats the proposals into the
  shape rendered in the ValueError message

utils/find_optimal_configs.py imports them from aetherscan.cli now and
becomes a thin CLI wrapper around the same logic. Dead lcm helpers
(unreferenced anywhere) dropped along the way. Defensive --num-replicas >= 1
check added to the script's main() since validate_args isn't on its path.
